+++ crates/rust-client/src/rpc/tonic_client/mod.rs
@@ -4,11 +4,11 @@ use alloc::{
     string::{String, ToString},
     vec::Vec,
 };
-use std::time::Duration;
+use std::{collections::BTreeMap, time::Duration};
 
 use async_trait::async_trait;
 use miden_objects::{
-    accounts::{Account, AccountId},
+    accounts::{Account, AccountCode, AccountId},
     crypto::merkle::{MerklePath, MmrProof},
     notes::{Note, NoteId, NoteInclusionProof, NoteTag, Nullifier},
     transaction::ProvenTransaction,
@@ -292,25 +292,18 @@ impl NodeRpcClient for TonicRpcClient {
     async fn get_account_proofs(
         &mut self,
         account_ids: &BTreeSet<AccountId>,
-        code_commitments: &[Digest],
+        known_account_codes: Vec<AccountCode>,
         include_headers: bool,
     ) -> Result<AccountProofs, RpcError> {
         let account_ids: Vec<AccountId> = account_ids.iter().copied().collect();
 
-        // Ensure all account IDs are public
-        for account_id in &account_ids {
-            if !account_id.is_public() {
-                return Err(RpcError::RequestError(
-                    NodeRpcClientEndpoint::GetAccountProofs.to_string(),
-                    format!("Account ID {account_id} is not public"),
-                ));
-            }
-        }
+        let known_account_codes: BTreeMap<Digest, AccountCode> =
+            known_account_codes.into_iter().map(|c| (c.commitment(), c)).collect();
 
         let request = GetAccountProofsRequest {
             account_ids: account_ids.iter().map(|&id| id.into()).collect(),
             include_headers: Some(include_headers),
-            code_commitments: code_commitments.iter().map(|c| c.into()).collect(),
+            code_commitments: known_account_codes.keys().map(|c| c.into()).collect(),
         };
 
         let rpc_api = self.rpc_api().await?;
@@ -349,12 +342,12 @@ impl NodeRpcClient for TonicRpcClient {
                 )));
             }
 
-            let headers = if include_headers {
+            let headers = if include_headers && account_id.is_public() {
                 Some(
                     account
                         .state_header
                         .ok_or(RpcError::ExpectedDataMissing("Account.StateHeader".to_string()))?
-                        .into_domain(account_id)?,
+                        .into_domain(account_id, &known_account_codes)?,
                 )
             } else {
                 None