+++ crates/rust-client/src/store/web_store/sync/mod.rs
@@ -12,8 +12,10 @@ use serde_wasm_bindgen::from_value;
 use wasm_bindgen_futures::*;
 
 use super::{
-    accounts::lock_account, chain_data::utils::serialize_chain_mmr_node,
-    notes::utils::apply_note_updates_tx, transactions::utils::update_account, WebStore,
+    accounts::{lock_account, utils::update_account},
+    chain_data::utils::serialize_chain_mmr_node,
+    notes::utils::apply_note_updates_tx,
+    WebStore,
 };
 use crate::{
     store::StoreError,