+++ crates/rust-client/src/transactions/mod.rs
@@ -11,20 +11,18 @@ use core::fmt::{self};
 
 pub use miden_lib::transaction::TransactionKernel;
 use miden_objects::{
-    accounts::{
-        Account, AccountCode, AccountDelta, AccountHeader, AccountId, AccountStorageHeader,
-        AccountType,
-    },
+    accounts::{Account, AccountCode, AccountDelta, AccountId, AccountType},
     assets::{Asset, NonFungibleAsset},
     crypto::merkle::MerklePath,
     notes::{Note, NoteDetails, NoteId, NoteTag},
     transaction::{InputNotes, TransactionArgs},
-    vm::AdviceInputs,
     AssetError, Digest, Felt, Word, ZERO,
 };
+use miden_tx::TransactionExecutor;
 pub use miden_tx::{
     LocalTransactionProver, ProvingOptions, TransactionProver, TransactionProverError,
 };
+pub use request::{ForeignAccount, ForeignAccountInputs};
 use script_builder::{AccountCapabilities, AccountInterface};
 use tracing::info;
 
@@ -75,6 +73,8 @@ impl TransactionResult {
         transaction: ExecutedTransaction,
         note_screener: NoteScreener,
         partial_notes: Vec<(NoteDetails, NoteTag)>,
+        current_block_num: u32,
+        current_timestamp: Option<u64>,
     ) -> Result<Self, ClientError> {
         let mut relevant_notes = vec![];
 
@@ -82,7 +82,17 @@ impl TransactionResult {
             let account_relevance = note_screener.check_relevance(note).await?;
 
             if !account_relevance.is_empty() {
-                relevant_notes.push(note.clone().into());
+                let metadata = *note.metadata();
+                relevant_notes.push(InputNoteRecord::new(
+                    note.into(),
+                    current_timestamp,
+                    ExpectedNoteState {
+                        metadata: Some(metadata),
+                        after_block_num: current_block_num,
+                        tag: Some(metadata.tag()),
+                    }
+                    .into(),
+                ));
             }
         }
 
@@ -93,7 +103,7 @@ impl TransactionResult {
                 None,
                 ExpectedNoteState {
                     metadata: None,
-                    after_block_num: 0,
+                    after_block_num: current_block_num,
                     tag: Some(*tag),
                 }
                 .into(),
@@ -353,17 +363,12 @@ impl<R: FeltRng> Client<R> {
         let tx_script = transaction_request
             .build_transaction_script(self.get_account_capabilities(account_id).await?)?;
 
-        // Inject foreign account data
-        let (foreign_data_advice_inputs, foreign_account_codes, fpi_block_num) =
-            self.get_foreign_account_inputs(transaction_request.foreign_accounts()).await?;
+        let foreign_accounts = transaction_request.foreign_accounts().clone();
+        let mut tx_args = transaction_request.into_transaction_args(tx_script);
 
-        let tx_args = transaction_request
-            .into_transaction_args(tx_script)
-            .with_advice_inputs(foreign_data_advice_inputs);
-
-        foreign_account_codes
-            .iter()
-            .for_each(|code| self.tx_executor.load_account_code(code));
+        // Inject state and code of foreign accounts
+        let fpi_block_num =
+            self.inject_foreign_account_inputs(foreign_accounts, &mut tx_args).await?;
 
         let block_num = if let Some(block_num) = fpi_block_num {
             block_num
@@ -398,7 +403,14 @@ impl<R: FeltRng> Client<R> {
 
         let screener = NoteScreener::new(self.store.clone());
 
-        TransactionResult::new(executed_transaction, screener, future_notes).await
+        TransactionResult::new(
+            executed_transaction,
+            screener,
+            future_notes,
+            self.get_sync_height().await?,
+            self.store.get_current_timestamp(),
+        )
+        .await
     }
 
     /// Proves the specified transaction using a local prover, submits it to the network, and saves
@@ -504,7 +516,11 @@ impl<R: FeltRng> Client<R> {
 
         let mut updated_input_notes = vec![];
         for mut input_note_record in consumed_notes {
-            if input_note_record.consumed_locally(account_id, transaction_id)? {
+            if input_note_record.consumed_locally(
+                account_id,
+                transaction_id,
+                self.store.get_current_timestamp(),
+            )? {
                 updated_input_notes.push(input_note_record);
             }
         }
@@ -705,64 +721,52 @@ impl<R: FeltRng> Client<R> {
         })
     }
 
-    /// Fetches foreign public account data as needed and returns advice inputs and account codes.
-    /// Additionally, it returns the block number pertinent to the foreign data.
+    /// Injects foreign account data inputs into `tx_args` (account proof, code commitment and
+    /// storage data). Additionally loads the account code into the transaction executor.
+    ///
+    /// For any [ForeignAccount::Public] in `foreing_accounts`, these pieces of data are retrieved
+    /// from the network. For any [ForeignAccount::Private] account, inner data is used.
     ///
     /// Account data is retrieved for the node's current chain tip, so we need to check whether we
     /// currently have the corresponding block header data. Otherwise, we additionally need to
     /// retrieve it.
-    async fn get_foreign_account_inputs(
+    async fn inject_foreign_account_inputs(
         &mut self,
-        account_ids: &BTreeSet<AccountId>,
-    ) -> Result<(AdviceInputs, Vec<AccountCode>, Option<u32>), ClientError> {
-        let mut advice_inputs = AdviceInputs::default();
-        let mut account_codes = Vec::new();
-
-        if account_ids.is_empty() {
-            return Ok((AdviceInputs::default(), vec![], None));
+        foreign_accounts: BTreeSet<ForeignAccount>,
+        tx_args: &mut TransactionArgs,
+    ) -> Result<Option<u32>, ClientError> {
+        if foreign_accounts.is_empty() {
+            return Ok(None);
         }
 
-        // Get tracked commitments
-        let tracked_code = self
-            .store
-            .get_foreign_account_code(account_ids.iter().cloned().collect())
-            .await?;
+        let account_ids = foreign_accounts.iter().map(|acc| acc.account_id());
+        let known_account_codes =
+            self.store.get_foreign_account_code(account_ids.collect()).await?;
 
-        let tracked_commitments: Vec<Digest> =
-            tracked_code.values().map(|code| code.commitment()).collect();
+        let known_account_codes: Vec<AccountCode> = known_account_codes.into_values().collect();
 
-        // Fetch account proofs
+        // Fetch FPI account data
         let (block_num, fpi_data_vec) =
-            self.rpc_api.get_account_fpi_data(account_ids, &tracked_commitments).await?;
+            self.rpc_api.get_fpi_account_data(foreign_accounts, known_account_codes).await?;
 
-        for fpi_data in fpi_data_vec.into_iter() {
-            let account_header = fpi_data.account_header();
-            let account_code = match fpi_data.code() {
-                Some(account_code) => {
-                    self.store
-                        .upsert_foreign_account_code(account_header.id(), account_code.clone())
-                        .await?;
+        for fpi_account_data in fpi_data_vec {
+            let (account_id, merkle_path, foreign_account_inputs) = fpi_account_data.into_parts();
 
-                    account_code
-                },
-                None => tracked_code
-                    .get(&account_header.id())
-                    .expect("Account code should be tracked if it's not in the RPC response"),
-            };
+            if account_id.is_public() {
+                // Update our foreign account code cache
+                self.store
+                    .upsert_foreign_account_code(
+                        account_id,
+                        foreign_account_inputs.account_code().clone(),
+                    )
+                    .await?;
+            }
 
-            let storage_header = fpi_data.storage_header();
-
-            account_codes.push(account_code.clone());
-
-            let merkle_path = fpi_data.merkle_proof();
-
-            // Extend advice inputs using the extracted data
             extend_advice_inputs_for_account(
-                &mut advice_inputs,
-                account_header,
-                account_code,
-                storage_header,
-                merkle_path,
+                tx_args,
+                &mut self.tx_executor,
+                foreign_account_inputs,
+                &merkle_path,
             )?;
         }
 
@@ -778,7 +782,7 @@ impl<R: FeltRng> Client<R> {
             }
         }
 
-        Ok((advice_inputs, account_codes, Some(block_num)))
+        Ok(Some(block_num))
     }
 }
 
@@ -809,14 +813,16 @@ impl<R: FeltRng> Client<R> {
     }
 }
 
-/// Extends the advice inputs with account data and Merkle proofs.
+/// Extends the advice inputs with account data and Merkle proofs, and loads the necessary
+/// [code](AccountCode) in `tx_executor`.
 fn extend_advice_inputs_for_account(
-    advice_inputs: &mut AdviceInputs,
-    account_header: &AccountHeader,
-    account_code: &AccountCode,
-    storage_header: &AccountStorageHeader,
+    tx_args: &mut TransactionArgs,
+    tx_executor: &mut TransactionExecutor,
+    foreign_account_inputs: ForeignAccountInputs,
     merkle_path: &MerklePath,
 ) -> Result<(), ClientError> {
+    let (account_header, storage_header, account_code) = foreign_account_inputs.into_parts();
+
     let account_id = account_header.id();
     let account_nonce = account_header.nonce();
     let vault_root = account_header.vault_root();
@@ -836,7 +842,7 @@ fn extend_advice_inputs_for_account(
     }
 
     // Extend the advice inputs with the new data
-    advice_inputs.extend_map([
+    tx_args.extend_advice_map([
         // ACCOUNT_ID -> [ID_AND_NONCE, VAULT_ROOT, STORAGE_ROOT, CODE_ROOT]
         (
             foreign_id_root,
@@ -855,8 +861,9 @@ fn extend_advice_inputs_for_account(
     ]);
 
     // Extend the advice inputs with Merkle store data
-    advice_inputs
-        .extend_merkle_store(merkle_path.inner_nodes(account_id.into(), account_header.hash())?);
+    tx_args.extend_merkle_store(merkle_path.inner_nodes(account_id.into(), account_header.hash())?);
+
+    tx_executor.load_account_code(&account_code);
 
     Ok(())
 }
@@ -959,11 +966,14 @@ mod test {
             .unwrap();
 
         client
-            .import_account(AccountData::new(
-                account.clone(),
-                None,
-                miden_objects::accounts::AuthSecretKey::RpoFalcon512(secret_key.clone()),
-            ))
+            .import_account(
+                AccountData::new(
+                    account.clone(),
+                    None,
+                    miden_objects::accounts::AuthSecretKey::RpoFalcon512(secret_key.clone()),
+                ),
+                false,
+            )
             .await
             .unwrap();
         client.sync_state().await.unwrap();