+++ crates/web-client/src/import.rs
@@ -13,7 +13,7 @@ impl WebClient {
                 .map_err(|err| err.to_string())?;
             let account_id = account_data.account.id().to_string();
 
-            match client.import_account(account_data).await {
+            match client.import_account(account_data, false).await {
                 Ok(_) => {
                     let message = format!("Imported account with ID: {}", account_id);
                     Ok(JsValue::from_str(&message))